+++ src/main/resources/com/google/javascript/jscomp/js/mini_es6_runtime.js
@@ -1,0 +1,1 @@
+'require es6/util/makeiterator';