+++ src/main/scala/com/programmaticallyspeaking/ncd/transpile/ES5Transpiler.scala
@@ -1,0 +1,11 @@
+package com.programmaticallyspeaking.ncd.transpile
+
+trait ES5Transpiler {
+  /**
+    * Transpiles script input and returns ES5 code. This method is thread safe.
+    *
+    * @param input the script to transpile
+    * @return transpiled ES5 code
+    */
+  def transpile(input: String): String
+}